MEED-669: change top contributors to top challengers
