+++ portlets/src/main/resources/locale/portlet/Challenges_en.properties
@@ -196,7 +196,7 @@ gamification.overview.challengesOverviewTitle=Take a new challenge
 gamification.overview.programsOverviewTitle=Top contributions programs
 
 gamification.overview.programsOverviewSummary=Don't rush and {0}start checking these actions{1}
-gamification.overview.topContributorsSummary=Challenge yourself by taking actions and be known
+gamification.overview.topChallengersSummary=Challenge yourself by taking actions and be known
 gamification.overview.challengesOverviewSummary=Let's do smoothy by {0}doing this program{1}
 gamification.overview.reputationKudosSummary={0}Congratulate your peers{1} to encourage them to participate and to share more!
 gamification.overview.rewardsWalletSummary={0}Initialize your wallet{1} to earn tokens based on what you do in the platform.
