+++ portlets/src/main/webapp/html/topChallengers/index.html
@@ -1,0 +1,7 @@
+<div class="VuetifyApp">
+  <div id="topChallengers">
+    <script type="text/javascript">
+              require(['PORTLET/gamification-portlets/topChallengers'], app => app.init());
+      </script>
+  </div>
+</div>
