--- portlets/src/main/webapp/html/topContributors/index.html
@@ -1,7 +1,0 @@
-<div class="VuetifyApp">
-  <div id="topContributors">
-    <script type="text/javascript">
-              require(['PORTLET/gamification-portlets/topContributors'], app => app.init());
-      </script>
-  </div>
-</div>
